Fix caching mechanism inside CSSModules.loadPath
When `loadPath` is called the `seen` object is meant to capture the
absolute path as a property and then check to see if that prop exists
before trying again.

Before this the `seen` object was not updated until the promises were
all resolved which meant that the guard to prevent multiple checks was
never hit

This fixes the problem  by setting the key on the `seen` object to a
promise that resolves the result of `load`.